# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CornerLab API
  description: >
    Plataforma de inteligência estatística, backtesting e gestão de estratégia
    para o mercado de escanteios. A API nunca recomenda apostas — apenas
    organiza dados históricos, calcula estatísticas e simula cenários.
    O Módulo de Inteligência Estatística (grupo `intelligence`) segue as mesmas
    restrições: nenhum endpoint prevê resultados futuros, promete lucro ou
    recomenda mercados — todo indicador é derivado de dados históricos
    armazenados, com rastreabilidade (liga/temporada/jogos analisados/data de
    atualização) em `meta` nas respostas.
  version: "0.2.0"
servers:
  - url: /api/v1
...
paths:
...
  /filters/run:
    post:
      tags: [filters]
      summary: Executar filtro/backtest (Módulo 3)
      description: >
        Se um token JWT válido for enviado (opcional), a execução é registrada
        automaticamente no Histórico da Estratégia do usuário.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FilterRunRequest' }
      responses:
        '200': { description: Resultado do backtest }
  /filters:
    get:
      tags: [filters]
      summary: Listar filtros salvos (autenticado)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Lista de filtros salvos }
    post:
      tags: [filters]
      summary: Salvar filtro personalizado (autenticado)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SaveFilterRequest' }
      responses:
        '201': { description: Filtro salvo }
  /filters/{id}/duplicate:
    post:
      tags: [filters]
      summary: Duplicar filtro salvo (autenticado)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '201': { description: Filtro duplicado }
  /filters/{id}:
    delete:
      tags: [filters]
      summary: Excluir filtro salvo (autenticado)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Filtro excluído }
...
  /exports/dashboard:
    get:
      tags: [export]
      summary: Exportar Dashboard (CSV/Excel)
      parameters:
        - in: query
          name: team_id
          required: true
          schema: { type: integer }
        - in: query
          name: league_id
          schema: { type: integer }
        - in: query
          name: season_id
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
        - in: query
          name: format
          schema: { type: string, enum: [csv, xlsx], default: csv }
      responses:
        '200': { description: "Arquivo CSV ou XLSX" }
  /exports/comparator:
    get:
      tags: [export]
      summary: Exportar Comparador (CSV/Excel)
      parameters:
        - in: query
          name: team_a
          required: true
          schema: { type: integer }
        - in: query
          name: team_b
          required: true
          schema: { type: integer }
        - in: query
          name: league_id
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
        - in: query
          name: format
          schema: { type: string, enum: [csv, xlsx], default: csv }
      responses:
        '200': { description: "Arquivo CSV ou XLSX" }
  /exports/filters/run:
    get:
      tags: [export]
      summary: Exportar resultado do Simulador de Filtros (CSV/Excel)
      parameters:
        - in: query
          name: league_id
          required: true
          schema: { type: integer }
        - in: query
          name: season_ids
          schema: { type: string, description: "IDs separados por vírgula" }
        - in: query
          name: team_id
          schema: { type: integer }
        - in: query
          name: last_n_games
          schema: { type: integer }
        - in: query
          name: home_away
          schema: { type: string, enum: ['', home, away] }
        - in: query
          name: corners_threshold
          required: true
          schema: { type: integer }
        - in: query
          name: opponent_tier
          schema: { type: string }
        - in: query
          name: max_odds
          schema: { type: number }
        - in: query
          name: stake
          schema: { type: number }
        - in: query
          name: format
          schema: { type: string, enum: [csv, xlsx], default: csv }
      responses:
        '200': { description: "Arquivo CSV ou XLSX com todos os jogos do backtest" }
  /exports/ranking:
    get:
      tags: [export]
      summary: Exportar Ranking (CSV/Excel)
      parameters:
        - in: query
          name: league_id
          required: true
          schema: { type: integer }
        - in: query
          name: metric
          schema:
            type: string
            enum: [average_corners, consistency, growth, decline, conceded_most, conceded_least]
            default: average_corners
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
        - in: query
          name: top
          schema: { type: integer, default: 10 }
        - in: query
          name: format
          schema: { type: string, enum: [csv, xlsx], default: csv }
      responses:
        '200': { description: "Arquivo CSV ou XLSX" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FilterRunRequest:
      type: object
      required: [league_id, corners_threshold]
      properties:
        league_id: { type: integer }
        season_ids: { type: array, items: { type: integer } }
        team_id: { type: integer, nullable: true }
        last_n_games: { type: integer, example: 10 }
        home_away: { type: string, enum: ['', home, away] }
        corners_threshold: { type: integer, example: 5 }
        opponent_tier: { type: string, example: G6 }
        max_odds: { type: number, example: 1.6 }
        stake: { type: number, example: 1 }
    SaveFilterRequest:
      type: object
      required: [name, definition]
      properties:
        name: { type: string }
        description: { type: string }
        definition: { type: string, description: "JSON serializado do FilterRunRequest" }
...